Add an accessor for revisions for a Need
Pull the revisions out of the API response in the
Need model and make it available using attr_reader.

Before we assign it to the instance variable, turn
the hash into an OpenStruct using the method from
GdsApi::Response. Explicitly keep the changes field
as a hash, so we can do key/value iteration.