Send all Maslow-provided fields on a create or update request
When we create or update a need, make sure we send
all the fields which Maslow knows about with the
request. This is so that fields which have been
blanked by the user will still be sent in the
request to the Need API.

As this adds some extra cruft to the request stubs
in the tests, I've added a blank_need_request method
to return an empty need, on which present attributes
can be merged for individual test cases.