Display information about a need in NeedsController#show
No longer redirect the 'show' action to 'edit',
instead displaying some basic information about
the need and a link to the edit form.

This tidies up the needs controller test a bit and
updates the integration tests to update a need so
they can still navigate to the edit form.