Use a tweaked version of the need impact in the view mode
In the view mode, we want to use a slightly rephrased
form of the impact strings so that they fit into
the sentence "If GOV.UK didn't meet this need…".

To achieve this, a new helper method is added to the
need helper which converts the existing impact into
a key which is then looked up in the locale file.