Add met when, justifications, impact and evidence fields to the need
Use the Need's attributes to populate the remaining
bits of data in the need view. 

I've added a second need fixture response to assist
testing when there are empty values returned from
the Need API.

This involves adding some new need helper methods 
to format numbers and percentages in a useful way.