Add in_scope as a protected field in Maslow
This adds support for the "in_scope" field from the Need API. However,
because we want to add a separate action for this in the UI, this 
shouldn't be able to be changed by mass-assignment through the form.

This adds a new constant for protected fields, and reorganised some of 
the existing constants for fields to make it clearer how they work.

Protected fields:
- get read and write accessors
- are sent back to the Need API when the need is saved
- can't be assigned through the `update` method
- get assigned when an existing need is initialized in the same way that
read-only attributes are assigned