Convert blank strings to nil before sending to API
We currently send empty form fields as blank strings when we make a
request to the API. When we start to update other fields and save them
without mass-assignment, we're sending those same fields as nil instead.

We should make this at least somewhat consistent, so this commit changes
the as_json method so that blank strings are now converted to be nil
before we send them to the Need API.

This also updates the blank_need_request method in the test helper, and
a test which broke with this change.