Revert "This is obvious"
Depending on who is on the team, how familiar they are with Rails'
conventions and what else they're concentrating on at the time, what's
obvious to one person will not be obvious to another. The fact that this
has been added to the codebase suggests it is non-obvious enough to at
least one person to be worth noting.

More generally, I would suggest being extremely careful declaring
anything "obvious" or "simple". Consider the potential to upset someone
who doesn't know what is obvious or simple to you.

This reverts commit 4feb4f2184d4ad5883a2048630ed15aa1c991581.