Fix missing navbar on History page for dupe needs
Before this change, the navbar and need title disappeared
on the "History & Notes" tab for needs that were marked as dupes.

Git archeology seemed to suggest that this was originally done in order
to prevent editing of duplicate needs, but now the navbar itself contains
logic to hide the "Edit" option for duplicate needs, so this is no longer
necessary.