Stop need criteria being deleted on enter
Hitting return when changing fields in the needs form additional
results in the browser submtting the form using the first button in
the form. In this case, that is the delete button of the first
criteria, which results in the first criteria being deleted because
the delete_criteria param gets posted. This commit adds an additional
submit button that is hidden from the user just above the delete
buttons, preventing the delete inputs being submitted by mistake.