Use met when criteria from need search results
related Need API change to be merged first:
https://github.com/alphagov/govuk_need_api/pull/64

www.agileplannerapp.com/boards/173808/cards/5644

we're seeing time-outs while exporting even a
small set of needs as CSV. this is caused due
to Maslow calling Need API once for every need
row in the CSV only to fetch met when criteria.

given that met when criteria are an essential
piece of information without which editors can't
process the information they got in the CSV, we
should include these in the basic need result
set. this will serve all information needed by
Maslow to prepare a CSV in one API call.