Descoping a need is a special case of marking it invalid
This commit updates Maslow to expect (and send) the "not valid" status
instead of "out of scope" from the Need API, since marking a need out
of scope is just one of several reasons to mark a need as invalid.

Only the internals are changed here; the frontend and URLs will be updated
in subsequent commits.