Fix attempted persistence of dynamic attributes
Mongoid was borked in production as a consequence of bumping
alphagov/gds-sso to `11.0.0`. An attempt would be made to update the
given user with the `organisation_content_id` attribute but the `User`
model did not explicitly declare this attribute.

For reference, errors were of the following nature:

```
Mongoid::Errors::UnknownAttribute: Problem: Attempted to set a value for
'organisation_content_id' which is not allowed on the model User.
Summary: Without including Mongoid::Attributes::Dynamic in your model
and the attribute does not already exist in the attributes hash,
attempting to call User#organisation_content_id= for it is not allowed.
This is also triggered by passing the attribute to any method that
accepts an attributes hash, and is raised instead of getting a
NoMethodError. Resolution: You can include Mongoid::Attributes::Dynamic
if you expect to be writing values for undefined fields often.
```