Ensure SSO required attributes are present
This is in lieu of GDS-SSO's `User` module asserting on the presence of
its requisite fields upon inclusion. I'm going to make the
aforementioned change to GDS-SSO separately so for now this spec
guarantees the required attributes are present in this project's
datastore.

Also adds the missing `disabled` field declaration whose absence caused
issues upon deployment to production.