Remove search from Maslow
Maslow is running on a super old version of Elasticsearch (0.90). It's
also soon going to be the only thing running on that old Elasticsearch
cluster.

I've removed the search functionality because I suspect that nobody
is using it, and it's the most simple way to allow us to remove the
old Elasticsearch cluster and stop spending money on it.