Revert "Remove filtering by organisation"
And restore the functionality. This functionality was removed because
at the time it was believed not to be feasible to do with the
Publishing API without adding functionality there.

However, it turns out this was possible all along. This commit reverts
the removal of the functionality, as well as the removal of the
related tests, and adjusts the tests so that they still work and
pass.

This reverts commit 1293ae392f103df24b4ad3c34d6b0330dec9ac15 among
other things.