Move to use a Services module for API clients
A services module is the pattern used in more apps, rather than
mattr_accessors on the main application.  We also take the opportunity
to remove the api_client_credentials.rb file which is no longer used now
we have removed need API.  The deploy scripts have been updated to no
longer overwrite this file, and what we use from the support adapter
does not require a bearer token.